updated interfaces

+++ contracts/mock/MockedMessageRelayer.t.sol
@@ -5,11 +5,11 @@ import "../relayers/IMessageRelayer.sol";
 
 contract MockedMessageRelayer is IMessageRelayer {
 
-    function sendCrosschainMessage(uint256 destChain, address destAddress, string calldata _message) external override {
+    function sendCrosschainMessage(uint32 destChain, address destAddress, string calldata _message) external override {
         //emit MessageRelayed(sourceChainId, destinationChainId, sender, destination, payload);
 		}
 
-    function onCrosschainMessage(uint256 destChain, address destAddress, uint256 amount) external override {
+    function onCrosschainMessage(uint32 destChain, address destAddress, uint256 amount) external override {
 
 		}
