+++ contracts/relayers/IMessageRelayer.sol
@@ -7,8 +7,8 @@ interface IMessageRelayer is IRelayer{
 
 	event MessageRelayed(uint256 indexed sourceChainId, uint256 indexed destinationChainId, address indexed sender, address destination, bytes payload);
 
-	function sendCrosschainMessage(uint256 destChain, address destAddress, string calldata _message) external;
+	function sendCrosschainMessage(uint32 destChain, address destAddress, string calldata _message) external;
 
-	function onCrosschainMessage(uint256 destChain, address destAddress, uint256 amount) external;
+	function onCrosschainMessage(uint32 destChain, address destAddress, uint256 amount) external;
 
 }
